+++ src/commands/Init.js
@@ -1,46 +1,55 @@
-const chalk = require('chalk');
-const fs = require('fs');
+import chalk from 'chalk';
+import fs from 'fs';
+import { getPluginsFromConfig } from '../services/PluginConfiguration.js';
 
+/**
+ * Sanitize name to remove '-' and capitalize letter after.
+ * @param {string} name - Name to sanitize.
+ * @returns {string} Sanitized name.
+ */
 function sanitize(name) {
   return name.replaceAll(/(-.)/ig, (v) => v.replace('-', '').toUpperCase());
 }
 
-exports.setup = (program) => {
+/**
+ * Initialize plugin.
+ * @param {Object} program - Program to run command.
+ */
+export default function setup(program) {
   program.command('init')
     .description('Init plugin files')
     .action(async () => {
       let plugins = [];
 
-      if (fs.existsSync('plugin.config.json')) {
-        const pluginConfig = JSON.parse(fs.readFileSync('plugin.config.json', 'utf8'));
-        plugins = plugins.concat(pluginConfig.plugins);
-      } else {
-        console.log(`\n${chalk.red('✘')} plugin.config.json file was not found. Use 'cli install' to create it.\n`);
-        return;
-      }
+      const allPlugins = getPluginsFromConfig();
+      const packageJson = JSON.parse(fs.readFileSync('package.json', 'utf8'));
 
-      plugins.forEach((pluginName) => {
-        const srcDir = `./node_modules/${pluginName}/public/`;
-        const distDir = `./public/plugins/${pluginName}/`;
+      plugins = Object.keys(packageJson.dependencies)
+        .map((key) => allPlugins.find((plugin) => plugin.packageKey === key))
+        .filter((value) => !!value);
+
+      plugins.forEach((plugin) => {
+        const srcDir = `./node_modules/${plugin.packageKey}/public/`;
+        const distDir = `./public/plugins/${plugin.name}/`;
 
         if (!fs.existsSync(srcDir)) {
-          console.log(`\n${chalk.red('✘')} Plugin '${pluginName}': No icons and models are imported.`);
+          console.log(`\n${chalk.red('✘')} Plugin '${plugin.name}': No icons and models are imported.`);
           return;
         }
 
         fs.cpSync(srcDir, distDir, { recursive: true });
 
-        console.log(`\n${chalk.green('✔')} Plugin '${pluginName}': Icons and models are imported.`);
+        console.log(`\n${chalk.green('✔')} Plugin '${plugin.name}': Icons and models are imported.`);
       });
 
       if (plugins.length > 0) {
         const pluginImports = plugins.map(
-          (pluginName) => `import ${sanitize(pluginName)} from '${pluginName}';`,
+          (plugin) => `import ${sanitize(plugin.name)} from '${plugin.packageKey}';`,
         );
         const content = [
           pluginImports.join('\n'),
           '\nexport default {',
-          plugins.map((p) => `${sanitize(p)},`).join('\n'),
+          plugins.map((plugin) => `${sanitize(plugin.name)},`).join('\n'),
           '};',
         ].join('\n');
 
@@ -53,4 +62,4 @@ exports.setup = (program) => {
         console.log(`\n${chalk.green('✔')} 'src/plugins/index.js' has been created or updated.\n`);
       }
     });
-};
+}