+++ changelog.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [Unreleased]
+
+### Added
+
+- Add workflow to publish on npm.
+
+### Changed
+
+- Update nodejs to version `20.16.0`.
+- Update npm to version `10.8.1`.
+- All plugins must now be defined in the configuration file to be installed.
+- Command line options:
+  - Remove `repository-name` and `repository-url`.
+  - Add `name` and `version`.
+
 ## [1.0.0] - 2023/05/03
 
 ### Added